chnages made to deploy the code on vercel
