+++ api/index.py
@@ -1,7 +1,13 @@
+import requests
 from fastapi import FastAPI
 
 app = FastAPI()
 
-@app.get("/api/python")
-def hello_world():
-    return {"message": "Hello World"}
+@app.get("/api/recipe")
+def get_recipe(recipe_id: int, servings: int):
+    url = "https://dinnerfy-ds-production.up.railway.app/recipe"
+    payload = {"RecipeID": recipe_id, "Servings": servings}
+    response = requests.post(url, json=payload)
+    recipe_details = response.json()
+    return recipe_details
+
