+++ app/layout.tsx
@@ -1,3 +1,4 @@
+import RecipeProvider from '../context'
 import './globals.css'
 import { Inter } from 'next/font/google'
 
@@ -15,7 +16,9 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
+      <RecipeProvider>
       <body className={inter.className}>{children}</body>
+      </RecipeProvider>
     </html>
   )
 }
