+++ app/page.tsx
@@ -1,8 +1,6 @@
-import Image from "next/image";
-import Link from "next/link";
-import InputSearch from "./components/InputSerach";
-import Nav from "./components/Nav";
-import Banner from "./components/Banner";
+import InputSearch from "../components/InputSerach";
+import Nav from "../components/Nav";
+import Banner from "../components/Banner";
 
 export default function Home() {
   return (
