+++ package.json
@@ -11,7 +11,6 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "axios": "^1.4.0",
     "concurrently": "^8.0.1",
     "eslint": "8.41.0",
     "eslint-config-next": "13.4.4",
@@ -22,12 +21,12 @@
     "react-loading": "^2.0.3"
   },
   "devDependencies": {
-    "autoprefixer": "10.4.14",
-    "tailwindcss": "3.3.2",
-    "postcss": "8.4.23",
-    "typescript": "5.0.4",
     "@types/node": "20.2.4",
     "@types/react": "18.2.7",
-    "@types/react-dom": "18.2.4"
+    "@types/react-dom": "18.2.4",
+    "autoprefixer": "10.4.14",
+    "postcss": "8.4.23",
+    "tailwindcss": "3.3.2",
+    "typescript": "5.0.4"
   }
 }
